move tokenization to TransformersModelManager
